// Client/Window/angular/projects/main/src/app/components/lobbychoice/enums/weapon-hash.enum.ts
export enum WeaponHash {
    Sniperrifle = 100416529,
    Fireextinguisher = 101631238,
    Compactlauncher = 125959754,
    Snowball = 126349499,
    Vintagepistol = 137902532,
    Combatpdw = 171789620,
    Heavysniper_mk2 = 177293209,
    Heavysniper = 205991906,
    Autoshotgun = 317205821,
    Microsmg = 324215364,
    Wrench = 419712736,
    Pistol = 453432689,
    Pumpshotgun = 487013001,
    Appistol = 584646201,
    Ball = 600439132,
    Molotov = 615608432,
    CeramicPistol = 727643628,
    Smg = 736523883,
    Stickybomb = 741814745,
    Petrolcan = 883325847,
    Stungun = 911657153,
    Stone_hatchet = 940833800,
    Assaultrifle_mk2 = 961495388,
    Heavyshotgun = 984333226,
    Minigun = 1119849093,
    Golfclub = 1141786504,
    Raycarbine = 1198256469,
    Flaregun = 1198879012,
    Flare = 1233104067,
    Grenadelauncher_smoke = 1305664598,
    Hammer = 1317494643,
    Pumpshotgun_mk2 = 1432025498,
    Combatpistol = 1593441988,
    Gusenberg = 1627465347,
    Compactrifle = 1649403952,
    Hominglauncher = 1672152130,
    Nightstick = 1737195953,
    Marksmanrifle_mk2 = 1785463520,
    Railgun = 1834241177,
    Sawnoffshotgun = 2017895192,
    Smg_mk2 = 2024373456,
    Bullpuprifle = 2132975508,
    Firework = 2138347493,
    Combatmg = 2144741730,
    Carbinerifle = 2210333304,
    Crowbar = 2227010557,
    Bullpuprifle_mk2 = 2228681469,
    Snspistol_mk2 = 2285322324,
    Flashlight = 2343591895,
    Proximine = 2381443905,
    NavyRevolver = 2441047180,
    Dagger = 2460120199,
    Grenade = 2481070269,
    Poolcue = 2484171525,
    Bat = 2508868239,
    Specialcarbine_mk2 = 2526821735,
    Doubleaction = 2548703416,
    Pistol50 = 2578377531,
    Knife = 2578778090,
    Mg = 2634544996,
    Bullpupshotgun = 2640438543,
    Bzgas = 2694266206,
    Unarmed = 2725352035,
    Grenadelauncher = 2726580491,
    Musket = 2828843422,
    Advancedrifle = 2937143193,
    Raypistol = 2939590305,
    Rpg = 2982836145,
    Rayminigun = 3056410471,
    Pipebomb = 3125143736,
    HazardCan = 3126027122,
    Minismg = 3173288789,
    Snspistol = 3218215474,
    Pistol_mk2 = 3219281620,
    Assaultrifle = 3220176749,
    Specialcarbine = 3231910285,
    Revolver = 3249783761,
    Marksmanrifle = 3342088282,
    Revolver_mk2 = 3415619887,
    Battleaxe = 3441901897,
    Heavypistol = 3523564046,
    Knuckle = 3638508604,
    Machinepistol = 3675956304,
    Combatmg_mk2 = 3686625920,
    Marksmanpistol = 3696079510,
    Machete = 3713923289,
    Switchblade = 3756226112,
    Assaultshotgun = 3800352039,
    Dbshotgun = 4019527611,
    Assaultsmg = 4024951519,
    Hatchet = 4191993645,
    Bottle = 4192643659,
    Carbinerifle_mk2 = 4208062921,
    Parachute = 4222310262,
    Smokegrenade = 4256991824
}